requirements.txt
# Critical-path Python pipeline
pandas>=2.0
numpy>=1.24
ta>=0.11
scikit-learn>=1.4
lightgbm>=4.0
pyyaml>=6.0
